requirements.txt
python-social-auth==0.2.11
Django==1.4.20

